+++ sign.js
@@ -1,6 +1,6 @@
 // @ts-check
-import sha256 from "./sha256.js";
 import hmac_sha256 from "./hmac_sha256.js";
+import sha256 from "./sha256.js";
 import { get_signature } from "./utils.js";
 import { array_to_number } from "./utils.js";
 /**